updated the nodejs server structure
