--- package.json
@@ -1,37 +1,0 @@
-{
-  "name": "red-tetris",
-  "version": "1.0.0",
-  "description": "This project, Red Tetris, aims to build a fully functional online, real-time multiplayer Tetris game, developed entirely with Full Stack JavaScript.",
-  "main": "index.js",
-  "scripts": {
-    "dev:server": "nodemon server/index.js",
-    "dev:client": "cd client && npm run dev",
-    "dev": "concurrently \"npm run dev:server\" \"npm run dev:client\"",
-    "build": "cd client && npm run build",
-    "start": "node server/index.js",
-    "test": "jest --coverage"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/mehdiakacem/red-tetris.git"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "type": "module",
-  "bugs": {
-    "url": "https://github.com/mehdiakacem/red-tetris/issues"
-  },
-  "homepage": "https://github.com/mehdiakacem/red-tetris#readme",
-  "dependencies": {
-    "express": "^5.1.0",
-    "socket.io": "^4.8.1"
-  },
-  "devDependencies": {
-    "@testing-library/jest-dom": "^6.9.1",
-    "@testing-library/react": "^16.3.0",
-    "concurrently": "^9.2.1",
-    "jest": "^30.2.0",
-    "nodemon": "^3.1.11"
-  }
-}
